# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DIU QuestionBank API",
    "version": "1.0.0",
    "description": "Backend API for **DIU QuestionBank**, running on Cloudflare Workers.\n\n## Authentication\n\nSign in with a verified `@diu.edu.bd` Google account. Obtain a Google ID token client-side\n(Google Sign-In), then exchange it for a DIU QuestionBank JWT:\n\n```http\nPOST /auth/google\nContent-Type: application/json\n\n{ \"idToken\": \"<google-id-token>\" }\n```\n\nThe response includes a JWT — pass it on every authenticated request:\n\n```\nAuthorization: Bearer <token>\n```\n\nIf the verified DIU email already exists, the existing user is signed in (`200`). If not, a new\nuser is created with an opaque generated username (`201`). Its role is `\"admin\"` when the\nemail matches the configured `ADMIN_EMAIL`; otherwise it is `\"user\"`. Google accounts outside\nthe `diu.edu.bd` domain are rejected (`403`).\n\n## Conventions\n\n- All timestamps are **Unix epoch seconds** (UTC integers), not ISO 8601 strings.\n- Errors return `{ \"error\": \"<message>\" }`. Validation failures additionally include an\n  `issues` array with field-level details."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from `POST /auth/google`."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "issues": {
            "description": "Field-level details, present on validation failures.",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "field",
                "message"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "Department": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "name": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "shortName"
        ],
        "additionalProperties": false
      },
...
      "CreateDepartment": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "shortName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20
          }
        },
        "required": [
          "name",
          "shortName"
        ],
        "additionalProperties": false
      },
      "UpdateDepartment": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "shortName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20
          }
        },
        "additionalProperties": false
      },
...
      "DepartmentList": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "name": {
                  "type": "string"
                },
                "shortName": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "shortName"
              ],
              "additionalProperties": false
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "perPage": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "total": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "totalPages": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "page",
              "perPage",
              "total",
              "totalPages"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/admin/departments": {
      "get": {
        "tags": [
          "admin-departments"
        ],
        "summary": "List departments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-badges": [
          {
            "name": "Admin",
            "color": "#dc2626",
            "position": "after"
          }
        ],
        "description": "**Access:** `Admin` — Requires a bearer token from an account with `role: \"admin\"`.\n\nPaginated list of departments.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "1-based page number."
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Items per page (max 100)."
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Case-insensitive substring match on name or short name."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentList"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-departments"
        ],
        "summary": "Create a department",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-badges": [
          {
            "name": "Admin",
            "color": "#dc2626",
            "position": "after"
          }
        ],
        "description": "**Access:** `Admin` — Requires a bearer token from an account with `role: \"admin\"`.\n\nCreates a new department.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartment"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "A conflicting record already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/departments/{id}": {
      "get": {
        "tags": [
          "admin-departments"
        ],
        "summary": "Get a department by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-badges": [
          {
            "name": "Admin",
            "color": "#dc2626",
            "position": "after"
          }
        ],
        "description": "**Access:** `Admin` — Requires a bearer token from an account with `role: \"admin\"`.\n\nA single department.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "department id."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin-departments"
        ],
        "summary": "Update a department",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-badges": [
          {
            "name": "Admin",
            "color": "#dc2626",
            "position": "after"
          }
        ],
        "description": "**Access:** `Admin` — Requires a bearer token from an account with `role: \"admin\"`.\n\nUpdates a department. Only the fields you send are changed.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "department id."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "A conflicting record already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-departments"
        ],
        "summary": "Delete a department",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-badges": [
          {
            "name": "Admin",
            "color": "#dc2626",
            "position": "after"
          }
        ],
        "description": "**Access:** `Admin` — Requires a bearer token from an account with `role: \"admin\"`.\n\nDeletes a department. Blocked with 409 if other records reference it.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "department id."
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted — no content"
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "In use by other records",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}